modified code style means divided to each class into separate package
