+++ src/com/streams/StreamsAllExamples.java
@@ -1,6 +1,5 @@
 package com.streams;
 
-import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.DoubleSummaryStatistics;
 import java.util.List;
@@ -8,147 +7,69 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
 
+import com.streams.employee.data.loadEmployeesData;
+import com.streams.entity.Employee;
+
 public class StreamsAllExamples {
 
 	public static void main(String[] args) {
 
-		List<String> listOf = loadEmployees().stream().filter(i -> i.id == 133)
+		List<String> listOf = loadEmployeesData.loadEmployees().stream().filter(i -> i.id == 133)
 				.map(i -> i.getName() + "--" + i.getSalary()).collect(Collectors.toList());
 		System.out.println(listOf);
 
-		loadEmployees().stream().limit(4).distinct().map(i -> i.getName() + "--" + i.getId() + "--" + i.getAge())
-				.collect(Collectors.toList()).forEach(i -> System.out.println(i));
+		loadEmployeesData.loadEmployees().stream().limit(4).distinct()
+				.map(i -> i.getName() + "--" + i.getId() + "--" + i.getAge()).collect(Collectors.toList())
+				.forEach(i -> System.out.println(i));
 
 		// max salary
-		loadEmployees().stream().collect(Collectors.maxBy(Comparator.comparing(i -> i.getSalary())))
+		loadEmployeesData.loadEmployees().stream().collect(Collectors.maxBy(Comparator.comparing(i -> i.getSalary())))
 				.map(i -> i.getName() + "--" + i.getSalary()).ifPresent(i -> System.out.println("max -" + i));
 
 		// min salary
-		loadEmployees().stream().collect(Collectors.minBy(Comparator.comparing(i -> i.salary)))
+		loadEmployeesData.loadEmployees().stream().collect(Collectors.minBy(Comparator.comparing(i -> i.salary)))
 				.map(i -> i.getName() + "--" + i.getSalary()).ifPresent(i -> System.out.println("min salary" + i));
 
 		// Average of salary
-		Double averageSalary = loadEmployees().stream().collect(Collectors.averagingDouble(i -> i.salary));
+		Double averageSalary = loadEmployeesData.loadEmployees().stream()
+				.collect(Collectors.averagingDouble(i -> i.salary));
 		System.out.println("averageSalary -- " + averageSalary);
 
 		// sum of salary
-		DoubleSummaryStatistics summ = loadEmployees().stream()
+		DoubleSummaryStatistics summ = loadEmployeesData.loadEmployees().stream()
 				.collect(Collectors.summarizingDouble(i -> i.salary + 2));
 		System.out.println("sum -" + summ.getSum());
 
 		// how many males and female employees
-		Map<String, Long> gender = loadEmployees().stream()
+		Map<String, Long> gender = loadEmployeesData.loadEmployees().stream()
 				.collect(Collectors.groupingBy(Employee::getGender, Collectors.counting()));
 		System.out.println(gender);
 
 		// how many departments
-		Map<String, Long> departmentList = loadEmployees().stream()
+		Map<String, Long> departmentList = loadEmployeesData.loadEmployees().stream()
 				.collect(Collectors.groupingBy(Employee::getDepartment, Collectors.counting()));
+		
 		for (Entry<String, Long> mapl : departmentList.entrySet()) {
 			System.out.println("Key: " + mapl.getKey() + ", Value: " + mapl.getValue());
-
 		}
 
 		// get only departmentNames
-		loadEmployees().stream().map(Employee::getDepartment).distinct().forEach(i -> System.out.println(i));
+		loadEmployeesData.loadEmployees().stream().map(Employee::getDepartment).distinct()
+				.forEach(i -> System.out.println(i));
 
 		// get the age of employees without duplicate age
-		loadEmployees().stream().map(e -> e.age).distinct().forEach(e -> System.out.println(e));
+		loadEmployeesData.loadEmployees().stream().map(e -> e.age).distinct().forEach(e -> System.out.println(e));
 
 		// What is the average age of male and female employees?
-		Map<String, Double> averageofAge = loadEmployees().stream()
+		Map<String, Double> averageofAge = loadEmployeesData.loadEmployees().stream()
 				.collect(Collectors.groupingBy(Employee::getGender, Collectors.averagingDouble(Employee::getAge)));
 		System.out.println("averageofAge-- " + averageofAge);
 
 		// What is the average age of male and female employees?
-		Map<String, Double> averageOfSalary = loadEmployees().stream()
+		Map<String, Double> averageOfSalary = loadEmployeesData.loadEmployees().stream()
 				.collect(Collectors.groupingBy(Employee::getGender, Collectors.averagingDouble(Employee::getSalary)));
-		System.out.println("averageOfSalary - "+averageOfSalary);
+		System.out.println("averageOfSalary - " + averageOfSalary);
 
-	}
-
-	private static List<Employee> loadEmployees() {
-		List<Employee> employeeList = new ArrayList<Employee>();
-
-		employeeList.add(new Employee(111, "Jiya Brein", 32, "Female", "HR", 2011, 25000.0));
-		employeeList.add(new Employee(122, "Paul Niksui", 25, "Male", "Sales And Marketing", 2015, 13500.0));
-		employeeList.add(new Employee(133, "Martin Theron", 29, "Male", "Infrastructure", 2012, 18000.0));
-		employeeList.add(new Employee(144, "Murali Gowda", 28, "Male", "Product Development", 2014, 32500.0));
-		employeeList.add(new Employee(155, "Nima Roy", 27, "Female", "HR", 2013, 22700.0));
-		employeeList.add(new Employee(166, "Iqbal Hussain", 43, "Male", "Security And Transport", 2016, 10500.0));
-		employeeList.add(new Employee(177, "Manu Sharma", 35, "Male", "Account And Finance", 2010, 27000.0));
-		employeeList.add(new Employee(188, "Wang Liu", 31, "Male", "Product Development", 2015, 34500.0));
-		employeeList.add(new Employee(199, "Amelia Zoe", 24, "Female", "Sales And Marketing", 2016, 11500.0));
-		employeeList.add(new Employee(200, "Jaden Dough", 38, "Male", "Security And Transport", 2015, 11000.5));
-		employeeList.add(new Employee(211, "Jasna Kaur", 27, "Female", "Infrastructure", 2014, 15700.0));
-		employeeList.add(new Employee(222, "Nitin Joshi", 25, "Male", "Product Development", 2016, 28200.0));
-		employeeList.add(new Employee(233, "Jyothi Reddy", 27, "Female", "Account And Finance", 2013, 21300.0));
-		employeeList.add(new Employee(244, "Nicolus Den", 24, "Male", "Sales And Marketing", 2017, 10700.5));
-		employeeList.add(new Employee(255, "Ali Baig", 23, "Male", "Infrastructure", 2018, 12700.0));
-		employeeList.add(new Employee(266, "Sanvi Pandey", 26, "Female", "Product Development", 2015, 28900.0));
-		employeeList.add(new Employee(277, "Anuj Chettiar", 31, "Male", "Product Development", 2012, 35700.0));
-		return employeeList;
-
-	}
-}
-
-class Employee {
-	int id;
-
-	String name;
-
-	int age;
-
-	String gender;
-
-	String department;
-
-	int yearOfJoining;
-
-	double salary;
-
-	public Employee(int id, String name, int age, String gender, String department, int yearOfJoining, double salary) {
-		this.id = id;
-		this.name = name;
-		this.age = age;
-		this.gender = gender;
-		this.department = department;
-		this.yearOfJoining = yearOfJoining;
-		this.salary = salary;
-	}
-
-	public int getId() {
-		return id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public int getAge() {
-		return age;
-	}
-
-	public String getGender() {
-		return gender;
-	}
-
-	public String getDepartment() {
-		return department;
-	}
-
-	public int getYearOfJoining() {
-		return yearOfJoining;
-	}
-
-	public double getSalary() {
-		return salary;
-	}
-
-	@Override
-	public String toString() {
-		return "Id : " + id + ", Name : " + name + ", age : " + age + ", Gender : " + gender + ", Department : "
-				+ department + ", Year Of Joining : " + yearOfJoining + ", Salary : " + salary;
 	}
 
 }
