+++ src/com/streams/entity/Employee.java
@@ -1,0 +1,63 @@
+package com.streams.entity;
+
+public class Employee {
+	public int id;
+
+	public String name;
+
+	public int age;
+
+	public String gender;
+
+	public String department;
+
+	public int yearOfJoining;
+
+	public double salary;
+
+	public Employee(int id, String name, int age, String gender, String department, int yearOfJoining, double salary) {
+		this.id = id;
+		this.name = name;
+		this.age = age;
+		this.gender = gender;
+		this.department = department;
+		this.yearOfJoining = yearOfJoining;
+		this.salary = salary;
+	}
+
+	public int getId() {
+		return id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public int getAge() {
+		return age;
+	}
+
+	public String getGender() {
+		return gender;
+	}
+
+	public String getDepartment() {
+		return department;
+	}
+
+	public int getYearOfJoining() {
+		return yearOfJoining;
+	}
+
+	public double getSalary() {
+		return salary;
+	}
+
+	@Override
+	public String toString() {
+		return "Id : " + id + ", Name : " + name + ", age : " + age + ", Gender : " + gender + ", Department : "
+				+ department + ", Year Of Joining : " + yearOfJoining + ", Salary : " + salary;
+	}
+
+}
+
